// src/api/models/domain_ReferenceTrainerSessionInput.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { domain_ReferenceTrainerSessionUpdate } from './domain_ReferenceTrainerSessionUpdate';
export type domain_ReferenceTrainerSessionInput = {
    sessionTrack?: string;
    updates?: Array<domain_ReferenceTrainerSessionUpdate>;
};