// src/api/models/domain_ReferenceTrainerSessionUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type domain_ReferenceTrainerSessionUpdate = {
    externalVerseId?: string;
    /**
     * correct_first, correct_retry, wrong
     */
    outcome?: string;
    /**
     * reference, incipit, ending, context
     */
    track?: string;
};